src/zerocomp/requirements.txt
urllib3
torchinfo
numpy
tqdm
pillow
opencv-python
OpenEXR
omegaconf
hydra-core
transformers
bitsandbytes
timm
comet_ml
h5py
json5
accelerate
matplotlib
lightning
open3d
huggingface_hub
kornia
skylibs
diffusers==0.32.2
gradio